requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comtypes==1.4.11
cryptography==46.0.3
cssselect2==0.8.0
defusedxml==0.7.1
distro==1.9.0
fastapi==0.123.5
freetype-py==2.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.12.0
lxml==6.0.2
Markdown==3.10
numpy==2.2.6
openai==2.14.0
opencv-python==4.12.0.88
oscrypto==1.3.0
pillow==12.0.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.29.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
PyMuPDF==1.26.6
pyparsing==3.2.5
pypdf==6.4.0
pypiwin32==223
python-bidi==0.6.7
python-dotenv==1.2.1
python-multipart==0.0.20
pyttsx3==2.99
pywin32==310
PyYAML==6.0.3
reportlab==4.4.5
requests==2.32.5
rlPyCairo==0.4.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
svglib==1.6.0
tinycss2==1.5.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
uritools==5.0.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
webencodings==0.5.1
websockets==15.0.1
xhtml2pdf==0.2.17
youtube-transcript-api==1.2.3
